Rename filter-panel.js to filter-utils.js
